Prevent UniqueId collisions within WeakDom's API (#288)
Closes #285, using Approach 1 as described in the issue. I ended up
having to touch `WeakDom::transfer` as well, because that method could
also create collisions.

It feels a bit icky to hard code the property name in these checks, but
it's *very* unlikely that we or Roblox will ever want to give it a
different name.

This PR additionally changes `WeakDom::insert` to accept a none `Ref`
for the parent (previously, it'd just fall over and panic). The reason
why is that I wanted to reuse the `UniqueId` checking code in
`WeakDom::new` (to uphold the invariant even when the prop is in the
root builder), but `WeakDom::insert` did not allow a none parent. The
only difference for insertions with none parent vs some parent is
whether the parent's list of children is extended, so it's pretty
minimal & I think it makes sense.